refacotor the file structure of marl_benchmark
