+++ baselines/marl_benchmark/README.md
@@ -3,19 +3,19 @@
 This directory contains the scenarios, training environment, and agents used in the CoRL20 paper: [SMARTS: Scalable Multi-Agent ReinforcementLearning Training School for Autonomous Driving](...).
 
 **Contents,**
-- `agents/`: YAML files and some RLlib-based policy implementations
-- `metrics/`: Class definition of metrics (default by a basic Metric class)
-- `networks/`: Custom network implementations
-  - `communicate.py`: Used for Networked agent learning
-- `scenarios/`: Contains three types of scenarios tested in the paper
-- `wrappers/`: Environment wrappers
-- `evaluate.py`: The evaluation program
-- `run.py`: Executes multi-agent training
+- `marl_benchmark/agents/`: YAML files and some RLlib-based policy implementations
+- `marl_benchmark/metrics/`: Class definition of metrics (default by a basic Metric class)
+- `marl_benchmark/networks/`: Custom network implementations
+  - `marl_benchmark/communicate.py`: Used for Networked agent learning
+- `marl_benchmark/scenarios/`: Contains three types of scenarios tested in the paper
+- `marl_benchmark/wrappers/`: Environment wrappers
+- `marl_benchmark/evaluate.py`: The evaluation program
+- `marl_benchmark/run.py`: Executes multi-agent training
 
 ## Setup
 ```bash
 # git clone ...
-cd <projec/baseline/marl_benchmark>
+cd <project/baseline/marl_benchmark>
 
 # setup virtual environment; presently at least Python 3.7 and higher is officially supported
 python3.7 -m venv .venv
@@ -34,7 +34,7 @@ pip install -e .
 To run the training procedure,
 
 ```bash
-# from baselines/marl_benchmark/
+# from baselines/marl_benchmark/marl_benchmark/
 $ python3.7 run.py <scenario> -f <config_file>
 # E.x. python3.7 run.py scenarios/intersections/4lane -f agents/ppo/baseline-lane-control.yaml
 ```
@@ -42,7 +42,7 @@ $ python3.7 run.py <scenario> -f <config_file>
 To run the evaluation procedure for multiple algorithms,
 
 ```bash
-# from baselines/marl_benchmark/
+# from baselines/marl_benchmark/marl_benchmark/
 $ python evaluate.py <scenario> -f <config_files>
 # E.x. python3.7  evaluate.py scenarios/intersections/4lane \
 #          -f agents/ppo/baseline-lane-control.yaml \
