+++ baselines/marl_benchmark/marl_benchmark/agents/__init__.py
@@ -23,11 +23,12 @@ from pathlib import Path
 
 import gym
 import yaml
-from baselines.marl_benchmark.metrics import basic_handler as metrics
-from baselines.marl_benchmark.utils import format
-from baselines.marl_benchmark.wrappers import rllib as rllib_wrappers
 
-from baselines.marl_benchmark import common
+from marl_benchmark import common
+from marl_benchmark.metrics import basic_handler as metrics
+from marl_benchmark.utils import format
+from marl_benchmark.wrappers import rllib as rllib_wrappers
+
 from smarts.core.agent_interface import (
     OGM,
     RGB,
