+++ baselines/marl_benchmark/marl_benchmark/agents/maddpg/baseline-lane-control.yaml
@@ -1,9 +1,9 @@
-name: maac-baseline-lane-controller
+name: maddpg-baseline-lane-controller
 
 agent:
   state:
     wrapper:
-      name: FrameStack
+      name: EarlyDone
       num_stack: 3
     features:
       goal_relative_pos: True
@@ -24,23 +24,12 @@ interface:
 
 policy:
   framework: rllib
-  config:
-    model:
-      custom_model_config:
-        critic_mode: global
-        agent_number: 4
-        action_model:
-          activation: relu
-          hiddens: [256, 128, 64]
-        value_model:
-          activation: relu
-          hiddens: [512, 256, 128]
   trainer:
-    path: baselines.marl_benchmark.agents.maac.tf_policy
-    name: CA2CTrainer
+    path: marl_benchmark.agents.maddpg.maddpg
+    name: MADDPGTrainer
 
 run:
-  checkpoint_freq: 4
+  checkpoint_freq: 40
   checkpoint_at_end: True
   max_failures: 1000
   resume: False
@@ -48,11 +37,11 @@ run:
   stop:
     time_total_s: 14400
   config:
-    log_level: WARN
+    log_level: ERROR
     num_workers: 1
     num_gpus: 0
     horizon: 1000
     # learning
-
-checkpoint:
-    ./log/results/run/4lane-4/CA2C_FrameStack_0_2021-01-25_17-07-05b3z4dh58/checkpoint_4/checkpoint-4
+    learning_starts: 1024
+    # buffer_size: 1024
+    train_batch_size: 1024
