+++ baselines/marl_benchmark/marl_benchmark/common.py
@@ -31,7 +31,7 @@ from ray.rllib.evaluation import MultiAgentEpisode, RolloutWorker
 from ray.rllib.policy import Policy
 
 from smarts.core.controllers import ActionSpaceType
-from smarts.core.plan import PositionalGoal
+from smarts.core.scenario import PositionalGoal
 from smarts.core.sensors import Observation
 from smarts.core.utils.math import vec_2d
