+++ baselines/marl_benchmark/marl_benchmark/scenarios/double_merge/nocross/scenario.py
