+++ baselines/marl_benchmark/marl_benchmark/wrappers/rllib/early_done.py
@@ -19,7 +19,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-from baselines.marl_benchmark.wrappers.rllib.frame_stack import FrameStack
+from marl_benchmark.wrappers.rllib.frame_stack import FrameStack
 
 
 class EarlyDone(FrameStack):
