+++ baselines/marl_benchmark/marl_benchmark/wrappers/rllib/simple.py
@@ -19,7 +19,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-from baselines.marl_benchmark.wrappers.rllib.wrapper import Wrapper
+from marl_benchmark.wrappers.rllib.wrapper import Wrapper
 
 
 class Simple(Wrapper):
